metajam: Easily Download Data and Metadata from 'DataONE'

A set of tools to foster the development of reproducible analytical workflow by simplifying the download of data and metadata from 'DataONE' (https://www.dataone.org) and easily importing this information into R.

See also

Author

Maintainer: Julien Brun julien.brun@alumni.duke.edu (ORCID)

Authors:

Irene Steves (ORCID) (https://github.com/isteves)

Mitchell Maier (ORCID)

Kristen Peach peach@nceas.ucsb.edu (ORCID)

Nicholas Lyon lyon@nceas.ucsb.edu (ORCID) (https://njlyon0.github.io/)

Other contributors:

Nathan Hwangbo nathanhwangbo@gmail.com (ORCID) [contributor]

Derek Strong dstrong@nceas.ucsb.edu (ORCID) [contributor]

Colin Smith colin.smith@wisc.edu (ORCID) [contributor]

Regents of the University of California [copyright holder]