Read data and metadata based on download_d1_data() file structure

Reads data along with metadata into your R environment based on download_d1_data() file structure.

read_d1_files(folder_path, fnc = "read_csv", ...)

Arguments

folder_path
(character) Path to a directory where data and metadata are located.

fnc
(character) Function to be used to read the data (default is readr::read_csv()).

...
Parameters to pass into the function specified in fnc.

Value

(list) Named list containing data and metadata as data frames.

See also

Examples

data_folder <- system.file("extdata", "test_data", package = "metajam")
soil_moist_data <- read_d1_files(data_folder)

#> Parsed with column specification:
#> cols(
#>   attributeName = col_character(),
#>   attributeDefinition = col_character(),
#>   formatString = col_character(),
#>   measurementScale = col_character(),
#>   unit = col_character(),
#>   numberType = col_character(),
#>   missingValueCode = col_character(),
#>   missingValueCodeExplanation = col_character(),
#>   domain = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   value = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   Date = col_character(),
#>   `Unmanipulated Moisture (cm3 cm-3)` = col_double(),
#>   `Unmanipulated Moisure (SE)` = col_double(),
#>   `Unburned Moisture (cm3 cm-3)` = col_double(),
#>   `Unburned Moisture (SE)` = col_double(),
#>   `Low Severity Moisture (cm3 cm-3)` = col_double(),
#>   `Low Severity Moisture (SE)` = col_double(),
#>   `Moderate Severity Moisture (cm3 cm-3)` = col_double(),
#>   `Moderate Severity Moisture (SE)` = col_double(),
#>   `High Severity Moisture (cm3 cm-3)` = col_double(),
#>   `High Severity Moisture (SE)` = col_double()
#> )

# You can specify the function you would like to use to read the file and pass parameters
soil_moist_data_skipped <- read_d1_files(data_folder, "read.csv",
                                         skip = 8, stringsAsFactors = FALSE)

#> Parsed with column specification:
#> cols(
#>   attributeName = col_character(),
#>   attributeDefinition = col_character(),
#>   formatString = col_character(),
#>   measurementScale = col_character(),
#>   unit = col_character(),
#>   numberType = col_character(),
#>   missingValueCode = col_character(),
#>   missingValueCodeExplanation = col_character(),
#>   domain = col_character()
#> )

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   value = col_character()
#> )